requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.6.0
pymongo==4.9.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
apscheduler==3.10.4
google-generativeai==0.3.2

